JASTrack writeRegParam: 71.2% -> 72.0% (retain full-width operation selector)

+++ src/JSystem/JAudio/JASystem/JASTrack.cpp
@@ -841,7 +841,7 @@ void TTrack::writeTimeParam(u8 param)
 void TTrack::writeRegParam(u8 param)
 {
 	u8 bVar9 = param & 0xC;
-	u8 bVar8 = param & 0x3;
+	u32 bVar8 = param & 0x3;
 
 	u32 r26;
